https://issues.jboss.org/browse/JBIDE-8757 code sync-up with eclipse.org/bpel fixes
git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@30666 a97e2381-89e5-4abb-bab3-167db6db766c

+++ bpel/plugins/org.eclipse.bpel.ui/src/org/eclipse/bpel/ui/actions/BPELPrintAction.java
@@ -53,7 +53,7 @@ public class BPELPrintAction extends PrintAction {
 		protected void setupPrinterGraphicsFor(Graphics graphics, IFigure figure) {
 			super.setupPrinterGraphicsFor(graphics, figure);
 			
-			double dpiScale = getPrinter().getDPI().x / Display.getCurrent().getDPI().x;
+			double dpiScale = (double) getPrinter().getDPI().x / Display.getCurrent().getDPI().x;
 			
 			// put the print region in display coordinates
 			org.eclipse.draw2d.geometry.Rectangle printableRegion = getPrintRegion();
@@ -92,7 +92,7 @@ public class BPELPrintAction extends PrintAction {
 			figure.paint(g);
 			
 			
-			double dpiScale = getPrinter().getDPI().x / Display.getCurrent().getDPI().x;
+			double dpiScale = (double) getPrinter().getDPI().x / Display.getCurrent().getDPI().x;
 			
 			// put the print region in display coordinates
 			region = getPrintRegion();
